server/requirements.txt
fastapi
uvicorn
asyncpg
aioredis==1.3.1
pyjwt
python-dotenv